:recycle: refac: 市区町村取得ロジック（set_area_id、extract_city_from_address）をconcernsに分割
